Add Spring Auto-Restart link
Somewhere in Slack was mentioned by @christherama
Auto-Restart feature, and this is simply DZONE article, that helps to implement that.

I also put a new link that I've recently found however not tested yet.

Anyway it is always good to know that Spring is capable of auto-restarting, if some student will take a look here